test(locationsHash): Write filter tests (#195)
* test(locationshash): adds tests

fix #64

* adds test docs
