+++ src/lib/locationsHash.ts
@@ -1,0 +1,9 @@
+const locationsHash: { [index: string] : string | undefined } = {
+  "/popular": "popular",
+  "/upvoted": "upvoted",
+  "/discussed": "discussed",
+  "/recent": "recent",
+  "/myVotes": "myVotes",
+}
+
+export default locationsHash;
